requirements.txt

timm
torch>=2.0
torchvision
tqdm
pyyaml
